refactor(tracing): rename core.py to tracing_core.py and update imports
