+++ tests/tracing/test_tracing.py
@@ -7,8 +7,8 @@ from machine_data_model.tracing import (
     clear_traces,
     TraceEventType,
     TraceLevel,
+    export_traces_json,
 )
-from machine_data_model.tracing.core import export_traces_json
 
 
 class TestDataModelTracing:
@@ -88,19 +88,16 @@ class TestDataModelTracing:
     def test_tracing_records_method_calls(self) -> None:
         clear_traces()
         data_model = DataModel(trace_level=TraceLevel.METHODS)
-        
+
         # Create return value node
         return_var = NumericalVariableNode(id="return", name="return", value=0)
-        
+
         # Create a simple method that returns a value
         def test_callback(x: int = 5) -> int:
             return x * 2
-        
+
         method = MethodNode(
-            id="test_method", 
-            name="test", 
-            returns=[return_var],
-            callback=test_callback
+            id="test_method", name="test", returns=[return_var], callback=test_callback
         )
         data_model.root.add_child(method)
         data_model._register_nodes(data_model.root)
@@ -110,22 +107,22 @@ class TestDataModelTracing:
         collector = get_global_collector()
         method_start_events = collector.get_events(TraceEventType.METHOD_START)
         method_end_events = collector.get_events(TraceEventType.METHOD_END)
-        
+
         assert len(method_start_events) == 1
         assert len(method_end_events) == 1
-        
+
         start_event = method_start_events[0]
         end_event = method_end_events[0]
-        
+
         assert start_event.details["method_id"] == "test_method"
         assert start_event.details["args"] == {}
         assert isinstance(start_event.timestamp, float)
-        
+
         assert end_event.details["method_id"] == "test_method"
         assert end_event.details["returns"] == {"return": 10}  # x=5 * 2 = 10
         assert isinstance(end_event.details["execution_time"], float)
         assert end_event.details["execution_time"] > 0
         assert isinstance(end_event.timestamp, float)
-        
+
         # Verify the method actually executed correctly
         assert result.return_values == {"return": 10}
